Module mithril_client::aggregator_client

Mechanisms to exchange data with an Aggregator.

The AggregatorClient trait abstracts how the communication with an Aggregator is done. The clients that need to communicate only need to define their request using the AggregatorRequest enum.

An implementation using HTTP is available: AggregatorHTTPClient.

Structs

AggregatorHTTPClient

Responsible for HTTP transport and API version check.

Enums

AggregatorClientError

Error tied with the Aggregator client

AggregatorRequest

What can be read from an AggregatorClient.

Traits

AggregatorClient

API that defines a client for the Aggregator