mithril_client

Module cardano_block_client

A client to retrieve from an aggregator cryptographic proofs of membership for a subset of Cardano blocks.

In order to do so it defines a CardanoBlockClient which exposes the following features:

get_proof: get a cryptographic proof that the blocks with given hash are included in the global Cardano blocks set.

get: get a Cardano block snapshot data from its hash.

list: get the list of the latest available Cardano block snapshot.

Important: Verifying a proof only means that its cryptography is valid, in order to certify that a Cardano blocks subset is valid, the associated proof must be tied to a valid Mithril certificate (see the example below).

Get and verify Cardano block proof

To get and verify a Cardano block proof using the ClientBuilder.

#[cfg(feature = "unstable")]
use mithril_client::{ClientBuilder, MessageBuilder};

let client = ClientBuilder::aggregator("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").build()?;

// 1 - Get a proof from the aggregator and verify it
let cardano_block_proof = client.cardano_block().get_proof(&["block-1", "block-2"]).await?;
println!("Mithril could not certify the following blocks : {:?}", &cardano_block_proof.non_certified_blocks);

let verified_blocks = cardano_block_proof.verify()?;

// 2 - Verify its associated certificate chain
let certificate = client.certificate().verify_chain(&cardano_block_proof.certificate_hash).await?;

// 3 - Ensure that the proof is indeed signed in the associated certificate
let message = MessageBuilder::new().compute_cardano_blocks_proofs_message(&certificate, &verified_blocks);
if certificate.match_message(&message) {
    // All green, Mithril certifies that those blocks are part of the Cardano blocks set.
    println!("Certified blocks : {:?}", verified_blocks.certified_blocks());
}

Get a Cardano block snapshot

To get a Cardano block snapshot using the ClientBuilder.

 #[cfg(feature = "unstable")]
use mithril_client::ClientBuilder;

let client = ClientBuilder::aggregator("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").build()?;
let cardano_block_snapshot = client.cardano_block().get_snapshot("CARDANO_BLOCK_SNAPSHOT_HASH").await?.unwrap();

println!("Cardano block snapshot hash={}, epoch={}, block_number_signed={}", cardano_block_snapshot.hash, cardano_block_snapshot.epoch, cardano_block_snapshot.block_number_signed);

List available Cardano block snapshots

To list latest available Cardano block snapshots using the ClientBuilder.

 #[cfg(feature = "unstable")]
use mithril_client::ClientBuilder;

let client = ClientBuilder::aggregator("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").build()?;
let cardano_block_snapshots = client.cardano_block().list_snapshots().await?;

for cardano_block_snapshot in cardano_block_snapshots {
    println!("Cardano block snapshot hash={}, epoch={}, block_number_signed={}", cardano_block_snapshot.hash, cardano_block_snapshot.epoch, cardano_block_snapshot.block_number_signed);
}

Structs

CardanoBlockClient

HTTP client for CardanoBlocksAPI from the aggregator

Traits

CardanoBlockAggregatorRequest

Define the requests against an aggregator related to Cardano blocks.