mithril_client
Module cardano_database_client

A client to retrieve Cardano databases data from an Aggregator.

In order to do so it defines a CardanoDatabaseClient which exposes the following features:

get: get a Cardano database data from its hash

list: get the list of available Cardano database

Get a Cardano database

To get a Cardano database using the ClientBuilder.

use mithril_client::ClientBuilder;

let client = ClientBuilder::aggregator("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").build()?;
let cardano_database = client.cardano_database().get("CARDANO_DATABASE_HASH").await?.unwrap();

println!(
    "Cardano database hash={}, merkle_root={}, immutable_file_number={:?}",
    cardano_database.hash,
    cardano_database.merkle_root,
    cardano_database.beacon.immutable_file_number
);

List available Cardano databases

To list available Cardano databases using the ClientBuilder.

use mithril_client::ClientBuilder;

let client = ClientBuilder::aggregator("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").build()?;
let cardano_databases = client.cardano_database().list().await?;

for cardano_database in cardano_databases {
    println!("Cardano database hash={}, immutable_file_number={}", cardano_database.hash, cardano_database.beacon.immutable_file_number);
}

Structs

CardanoDatabaseClient

HTTP client for CardanoDatabase API from the Aggregator