mithril_client::cardano_database_client

Trait CardanoDatabaseAggregatorRequest

pub trait CardanoDatabaseAggregatorRequest: Send + Sync {
    // Required methods
    fn list_latest<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = MithrilResult<Vec<CardanoDatabaseSnapshotListItem>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn list_by_epoch<'life0, 'async_trait>(
        &'life0 self,
        specifier: EpochSpecifier,
    ) -> Pin<Box<dyn Future<Output = MithrilResult<Vec<CardanoDatabaseSnapshotListItem>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_by_hash<'life0, 'life1, 'async_trait>(
        &'life0 self,
        hash: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = MithrilResult<Option<CardanoDatabaseSnapshot>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn increment_cardano_database_complete_restoration_statistic<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = MithrilResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn increment_cardano_database_partial_restoration_statistic<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = MithrilResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn increment_immutables_snapshot_restored_statistic<'life0, 'async_trait>(
        &'life0 self,
        number_of_immutable_files_restored: u32,
    ) -> Pin<Box<dyn Future<Output = MithrilResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn increment_ancillary_downloaded_statistic<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = MithrilResult<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Define the requests against an aggregator related to Cardano database v2 snapshots.

Required Methods

fn list_latest<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = MithrilResult<Vec<CardanoDatabaseSnapshotListItem>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the list of the latest Cardano database v2 snapshots from the aggregator.

fn list_by_epoch<'life0, 'async_trait>( &'life0 self, specifier: EpochSpecifier, ) -> Pin<Box<dyn Future<Output = MithrilResult<Vec<CardanoDatabaseSnapshotListItem>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the list of the latest Cardano database v2 snapshots for an EpochSpecifier from the aggregator.

fn get_by_hash<'life0, 'life1, 'async_trait>( &'life0 self, hash: &'life1 str, ) -> Pin<Box<dyn Future<Output = MithrilResult<Option<CardanoDatabaseSnapshot>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Get the details of a Cardano database v2 snapshot for a given hash from the aggregator.

fn increment_cardano_database_complete_restoration_statistic<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = MithrilResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Notify the aggregator that a complete Cardano database v2 restoration has been performed.

fn increment_cardano_database_partial_restoration_statistic<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = MithrilResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Notify the aggregator that a partial Cardano database v2 restoration has been performed.

fn increment_immutables_snapshot_restored_statistic<'life0, 'async_trait>( &'life0 self, number_of_immutable_files_restored: u32, ) -> Pin<Box<dyn Future<Output = MithrilResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Notify the aggregator that a given number of immutable files has been downloaded.

fn increment_ancillary_downloaded_statistic<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = MithrilResult<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Notify the aggregator that a Cardano database v2 ancillary file has been downloaded.