mithril_client

Module cardano_stake_distribution_client

A client to retrieve Cardano stake distributions data from an aggregator.

In order to do so it defines a CardanoStakeDistributionClient which exposes the following features:

get: get a Cardano stake distribution data from its hash

get_by_epoch: get a Cardano stake distribution data from its epoch

list: get the list of available Cardano stake distribution

Get a Cardano stake distribution

To get a Cardano stake distribution using the ClientBuilder.

use mithril_client::ClientBuilder;

let client = ClientBuilder::aggregator("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").build()?;
let cardano_stake_distribution = client.cardano_stake_distribution().get("CARDANO_STAKE_DISTRIBUTION_HASH").await?.unwrap();

println!(
    "Cardano stake distribution hash={}, epoch={}, stake_distribution={:?}",
    cardano_stake_distribution.hash,
    cardano_stake_distribution.epoch,
    cardano_stake_distribution.stake_distribution
);

List available Cardano stake distributions

To list available Cardano stake distributions using the ClientBuilder.

use mithril_client::ClientBuilder;

let client = ClientBuilder::aggregator("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").build()?;
let cardano_stake_distributions = client.cardano_stake_distribution().list().await?;

for cardano_stake_distribution in cardano_stake_distributions {
    println!("Cardano stake distribution hash={}, epoch={}", cardano_stake_distribution.hash, cardano_stake_distribution.epoch);
}

Get a Cardano stake distribution by epoch

To get a Cardano stake distribution by epoch using the ClientBuilder.

Note: The epoch represents the epoch at the end of which the Cardano stake distribution is computed by the Cardano node

use mithril_client::{ClientBuilder, common::Epoch};

let client = ClientBuilder::aggregator("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").build()?;
// For a specific epoch
let cardano_stake_distribution = client.cardano_stake_distribution().get_by_epoch(Epoch(500)).await?.unwrap();
// For the latest epoch known by the Mithril aggregator
let cardano_stake_distribution = client.cardano_stake_distribution().get_for_latest_epoch().await?.unwrap();
// For the latest epoch known by the Mithril aggregator with an offset
let cardano_stake_distribution = client.cardano_stake_distribution().get_for_latest_epoch_with_offset(4).await?.unwrap();

println!(
    "Cardano stake distribution hash={}, epoch={}, stake_distribution={:?}",
    cardano_stake_distribution.hash,
    cardano_stake_distribution.epoch,
    cardano_stake_distribution.stake_distribution
);

Structs

CardanoStakeDistributionClient

HTTP client for CardanoStakeDistribution API from the aggregator

Traits

CardanoStakeDistributionAggregatorRequest

Define the requests against an aggregator related to Cardano stake distribution.