Module mithril_client::common

mithril-common re-exports

Structs

CardanoDbBeacon

A point in the Cardano chain at which a Mithril certificate of the Cardano Database should be produced.

ChainPoint

The Cardano chain point which is used to identify a specific point in the Cardano chain.

Epoch

Epoch represents a Cardano epoch

ProtocolMessage

ProtocolMessage represents a message that is signed (or verified) by the Mithril protocol

ProtocolParameters

Protocol cryptographic parameters

Enums

CompressionAlgorithm

Compression algorithm for the snapshot archive artifacts.

ProtocolMessagePartKey

The key of a ProtocolMessage

Type Aliases

BlockHash

Hash of a Cardano Block

BlockNumber

BlockNumber is the block number of a Cardano transaction.

ImmutableFileNumber

ImmutableFileNumber represents the id of immutable files in the Cardano node database

SlotNumber

Cardano Slot number

TransactionHash

TransactionHash is the unique identifier of a cardano transaction.