Trait mithril_client::services::MithrilStakeDistributionService

pub trait MithrilStakeDistributionService {
    // Required methods
    fn list<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = StdResult<Vec<MithrilStakeDistributionListItemMessage>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn download<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        hash: &'life1 str,
        dirpath: &'life2 Path,
        genesis_verification_key: &'life3 str
    ) -> Pin<Box<dyn Future<Output = StdResult<PathBuf>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
}

Definition of the service responsible of Mithril Stake Distribution.

Required Methods

fn list<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = StdResult<Vec<MithrilStakeDistributionListItemMessage>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Return a list of the certified Mithril stake distributions

fn download<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, hash: &'life1 str, dirpath: &'life2 Path, genesis_verification_key: &'life3 str ) -> Pin<Box<dyn Future<Output = StdResult<PathBuf>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Download and verify the specified stake distribution