Struct mithril_client::MithrilStakeDistributionListItem

pub struct MithrilStakeDistributionListItem {
    pub epoch: Epoch,
    pub hash: String,
    pub certificate_hash: String,
    pub created_at: DateTime<Utc>,
}

Item to display Mithril Stake Distribution lines in a table.

Fields
epoch: Epoch

Epoch at which the Mithril Stake Distribution is created

hash: String

Hash of the Mithril Stake Distribution (different from the AVK).

certificate_hash: String

Hash of the associated certificate

created_at: DateTime<Utc>

Date and time at which the Mithril Stake Distribution was created