Crate mithril_client

Define everything necessary to list, download, and validate snapshots from a Mithril Aggregator.

To query an aggregator for snapshots & certificate use the services::SnapshotService.

Currently, client::Client exposes the following features:

client::Client::show_snapshot: call the snapshot service to get a snapshot message from a digest

client::Client::list_snapshots: call the snapshot service to get the list of available snapshots

client::Client::list_mithril_stake_distributions: call the mithril stake distribution service for the list of available mithril stake distributions

Below are some examples describing the use of the library’s functions from your own project:

client::Client::show_snapshot

use mithril_client::client::Client;
use mithril_client::common::StdResult;

#[tokio::main]
async fn main() -> StdResult<()> {
    let client = Client::new("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").await?;
    let snapshot = client.show_snapshot("5a1288f7164bec049f34e46002e939f4c609a0ddf86636fdc4180ea22342cab7").await?;

    println!("Snapshot id={}, size={}", snapshot.digest, snapshot.size);

    Ok(())
}

client::Client::list_snapshots

use mithril_client::client::Client;
use mithril_client::common::StdResult;

#[tokio::main]
async fn main() -> StdResult<()> {
    let client = Client::new("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").await?;
    let snapshots = client.list_snapshots().await?;

    for snapshot in snapshots {
        println!("Snapshot id={}, size={}", snapshot.digest, snapshot.size);
    }

    Ok(())
}

client::Client::list_mithril_stake_distributions

use mithril_client::client::Client;
use mithril_client::common::StdResult;

#[tokio::main]
async fn main() -> StdResult<()> {
    let client = Client::new("YOUR_AGGREGATOR_ENDPOINT", "YOUR_GENESIS_VERIFICATION_KEY").await?;
    let stake_distributions = client.list_mithril_stake_distributions().await?;

    for stake_distribution in stake_distributions {
        println!("Mithril Stake Distribution hash={}", stake_distribution.hash);
    }

    Ok(())
}

Modules

aggregator_client

Transport layers to the Aggregator nodes

client

Client module

common

mithril-common re-exports

dependencies

Dependency manager for the client

services

Services This module contains the different services tied with their bounded context.

utils

Utilities module This module contains tools needed mostly in services layers.

Structs

FromCertificateMessageAdapter

Adapter to convert [CertificateMessage] to [Certificate] instances

FromSnapshotMessageAdapter

Adapter to convert SnapshotMessage to [SignedEntity<Snapshot>] instances