Crate mithril_client

Mithril Network / Mithril Client

This is a work in progress :hammer_and_wrench:s

A Mithril Client is able to download, verify authenticity and restore a snapshot into a Cardano Node.

It allows user to bootstrap a Cardano Node in minutes instead of days.

This cli implements a MVP version of a Mithril Client.

Pre-requisites:

Install Rust

Install a correctly configured Rust toolchain (version 1.58.0+).

Download source code:

git clone https://github.com/input-output-hk/mithril

cd mithril-network/mithril-client

Development test and build:

make test

make help

make doc

make debug

Release build and run binary:

make run list

make build

./mithril-client --help

./mithril-client

Build and run Docker container:

cd ../../
docker build -t mithril/mithril-client -f mithril-network/mithril-client/Dockerfile .

docker run --rm --name='mithril-client' mithril/mithril-client

Modules

aggregator

aggregator_fake

client

entities

errors

Structs

Args

CLI args

Enums

Commands

CLI command list

Functions

main

pretty_print_error

Pretty print error