Crate mithril_client

Define everything necessary to list, download, and validate snapshots from a Mithril Aggregator.

To query an aggregator for snapshots & certificate use the AggregatorHTTPClient that implement the AggregatorHandler trait.

To list, download, and validate snapshots use the Runtime. You must initialize it by giving it a CertificateVerifier, a ProtocolGenesisVerifier and a AggregatorHandler, and a Digester implementations using the with_xxx methods.

Modules

commands

Command module This module holds the subcommands that can be used from the CLI.

Structs

AggregatorHTTPClient

AggregatorHTTPClient is a http client for an aggregator

Config

Client configuration

Runtime

Mithril client runtime

Enums

AggregatorHandlerError

AggregatorHandler related errors.

RuntimeError

Runtime related errors.

Traits

AggregatorHandler

AggregatorHandler represents a read interactor with an aggregator

Functions

convert_to_field_items

Convert Snapshot to SnapshotFieldItems routine