mithril_client

Module common

mithril-common re-exports

Modules

test

Test utilities.

Structs

AncillaryMessagePart

The message part that represents the locations of the Cardano database ancillary.

BlockNumber

BlockNumber is the block number of a Cardano transaction.

CardanoDbBeacon

A point in the Cardano chain at which a Mithril certificate of the Cardano Database should be produced.

ChainPoint

The Cardano chain point which is used to identify a specific point in the Cardano chain.

DigestsMessagePart

The message part that represents the locations of the Cardano database digests.

Epoch

Epoch represents a Cardano epoch

ImmutablesMessagePart

The message part that represents the locations of the Cardano database immutables.

MKProof

A Merkle proof

MithrilNetwork

Representation of a Mithril network

ProtocolMessage

ProtocolMessage represents a message that is signed (or verified) by the Mithril protocol

ProtocolParameters

Protocol cryptographic parameters

SlotNumber

Cardano Slot number

TemplateUri

TemplateUri represents an URI pattern used to build a file’s location

Enums

AggregateSignatureType

The type of STM aggregate signature.

AncillaryLocation

Locations of the ancillary files.

CardanoNetwork

The Cardano Network that is being targeted

CompressionAlgorithm

Compression algorithm for the snapshot archive artifacts.

DigestLocation

Locations of the immutable file digests.

EpochSpecifier

Represents the different ways to specify an epoch when querying the API.

ImmutablesLocation

Locations of the immutable files.

MultiFilesUri

MultiFilesUri represents a unique location uri for multiple files

ProtocolMessagePartKey

The key of a ProtocolMessage

SignedEntityType

The signed entity type that represents a type of data signed by the Mithril protocol Note: Each variant of this enum must be associated to an entry in the signed_entity_type table of the signer/aggregator nodes. The variant are identified by their discriminant (i.e. index in the enum), thus the modification of this type should only ever consist of appending new variants.

SignedEntityTypeDiscriminants

The discriminants of the SignedEntityType enum.

SupportedEra

The era that the software is running or will run

Type Aliases

BlockHash

Hash of a Cardano Block

ImmutableFileNumber

ImmutableFileNumber represents the id of immutable files in the Cardano node database

MagicId

Cardano Network magic identifier

StakeDistribution

StakeDistribution represents the stakes of multiple participants in the Cardano chain

TransactionHash

TransactionHash is the unique identifier of a cardano transaction.