Module mithril_client::common

mithril-common re-exports

Structs

Beacon

Beacon represents a point in the Cardano chain at which a Mithril certificate should be produced

Epoch

Epoch represents a Cardano epoch

ProtocolGenesisVerifier

A protocol Genesis Verifier that is responsible for verifying the Genesis Certificate

SnapshotListItemMessage

Message structure of a snapshot list item

SnapshotMessage

Message structure of a snapshot

Enums

CompressionAlgorithm

Compression algorithm for the snapshot archive artifacts.

Traits

CertificateVerifier

CertificateVerifier is the cryptographic engine in charge of verifying multi signatures and certificates

Type Definitions

MithrilStakeDistributionListMessage

Message structure of a Mithril Stake Distribution list

ProtocolGenesisVerificationKey

Wrapper of Ed25519:PublicKey.

SnapshotListMessage

Message structure of a snapshot list

StdError

Generic error type

StdResult

Generic result type